backend/requirements.txt

fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
pydantic[email]==2.5.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
openai==1.3.7
requests==2.31.0
python-dotenv==1.0.0
email-validator==2.1.0
twilio==8.10.0
sendgrid==6.10.0
pandas==2.1.4
numpy==1.25.2
